requirements.txt
# ==========================================
# Core Dependencies
# ==========================================
colorama>=0.4.6
discord-webhook>=1.3.0
discord.py-self>=1.9.2
matplotlib>=3.7.0
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
httpx>=0.24.0

# ==========================================
# GUI Dependencies
# ==========================================
ttkbootstrap>=1.10.1

# ==========================================
# Broker APIs
# ==========================================
pyetrade>=1.3.0
ib_async>=0.9.86

# ==========================================
# Market Data APIs
# ==========================================
thetadata>=0.1.0

# ==========================================
# Development Dependencies (optional)
# ==========================================
# Uncomment these for development
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.7.0
# pylint>=2.17.0
# flake8>=6.1.0
